Checking for ::Delayed rather than for ::Delayed::Job is better under some circumstances.
DelayedJob sets ::Job constant when it's gessing for backend during Rails
initialization rpocess.
https://github.com/collectiveidea/delayed_job/blob/v2.0/lib/delayed/railtie.rb#L7

Hovewer, it seems that DelayedPaperclip initializes options hash earlier in
that process.
https://github.com/jstorimer/delayed_paperclip/blob/master/lib/delayed_paperclip/railtie.rb#L8

Thus, ::Delayed::Job const is not set yet, so checking for ::Delayed seems to
be better.

In Rails 3 this can be fixed by using :after options for Railtie, but Rails 2
doesn't have that :(